refactor(api): migrate Organization to src/organizational-entities
