+++ api/lib/domain/validators/organization-with-tags-and-target-profiles-script.js
@@ -1,8 +1,8 @@
 import Joi from 'joi';
 
+import { Organization } from '../../../src/organizational-entities/domain/models/Organization.js';
 import { SUPPORTED_LOCALES } from '../../../src/shared/domain/constants.js';
 import { EntityValidationError } from '../../../src/shared/domain/errors.js';
-import { Organization } from '../../../src/shared/domain/models/Organization.js';
 import { Membership } from '../models/Membership.js';
 
 const supportedLocales = SUPPORTED_LOCALES.map((supportedLocale) => supportedLocale.toLocaleLowerCase());
