+++ api/lib/infrastructure/repositories/membership-repository.js
@@ -1,5 +1,5 @@
 import { knex } from '../../../db/knex-database-connection.js';
-import { Organization } from '../../../src/shared/domain/models/Organization.js';
+import { Organization } from '../../../src/organizational-entities/domain/models/Organization.js';
 import { User } from '../../../src/shared/domain/models/User.js';
 import * as knexUtils from '../../../src/shared/infrastructure/utils/knex-utils.js';
 import { MembershipCreationError, MembershipUpdateError, NotFoundError } from '../../domain/errors.js';
