+++ api/tests/integration/infrastructure/utils/bookshelf-to-domain-converter_test.js
@@ -7,7 +7,7 @@ import { BookshelfCampaignParticipation } from '../../../../lib/infrastructure/o
 import { BookshelfOrganization } from '../../../../lib/infrastructure/orm-models/Organization.js';
 import { BookshelfUser } from '../../../../lib/infrastructure/orm-models/User.js';
 import * as bookshelfToDomainConverter from '../../../../lib/infrastructure/utils/bookshelf-to-domain-converter.js';
-import { Organization } from '../../../../src/shared/domain/models/Organization.js';
+import { Organization } from '../../../../src/organizational-entities/domain/models/Organization.js';
 import { User } from '../../../../src/shared/domain/models/User.js';
 import { databaseBuilder, expect } from '../../../test-helper.js';
