+++ api/tests/shared/integration/infrastructure/repositories/user-repository_test.js
@@ -15,7 +15,7 @@ import { Membership } from '../../../../../lib/domain/models/Membership.js';
 import { UserDetailsForAdmin } from '../../../../../lib/domain/models/UserDetailsForAdmin.js';
 import { OrganizationLearnerForAdmin } from '../../../../../lib/domain/read-models/OrganizationLearnerForAdmin.js';
 import { DomainTransaction } from '../../../../../lib/infrastructure/DomainTransaction.js';
-import { Organization } from '../../../../../src/shared/domain/models/Organization.js';
+import { Organization } from '../../../../../src/organizational-entities/domain/models/Organization.js';
 import { User } from '../../../../../src/shared/domain/models/User.js';
 import * as userRepository from '../../../../../src/shared/infrastructure/repositories/user-repository.js';
 import { catchErr, databaseBuilder, expect, knex, sinon } from '../../../../test-helper.js';
