+++ api/tests/shared/prescriber-management/integration/infrasctructure/repositories/prescriber-repository_test.js
@@ -5,9 +5,9 @@ import { UserNotFoundError } from '../../../../../../lib/domain/errors.js';
 import { Membership } from '../../../../../../lib/domain/models/Membership.js';
 import { Tag } from '../../../../../../lib/domain/models/Tag.js';
 import { UserOrgaSettings } from '../../../../../../lib/domain/models/UserOrgaSettings.js';
+import { Organization } from '../../../../../../src/organizational-entities/domain/models/Organization.js';
 import { ORGANIZATION_FEATURE } from '../../../../../../src/shared/domain/constants.js';
 import { ForbiddenAccess } from '../../../../../../src/shared/domain/errors.js';
-import { Organization } from '../../../../../../src/shared/domain/models/Organization.js';
 import { Prescriber } from '../../../../../../src/shared/prescriber-management/domain/read-models/Prescriber.js';
 import * as prescriberRepository from '../../../../../../src/shared/prescriber-management/infrastructure/repositories/prescriber-repository.js';
 import { catchErr, databaseBuilder, expect } from '../../../../../test-helper.js';
