+++ api/tests/unit/infrastructure/serializers/jsonapi/organization-serializer_test.js
@@ -1,6 +1,6 @@
 import { Tag } from '../../../../../lib/domain/models/Tag.js';
 import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/organization-serializer.js';
-import { Organization } from '../../../../../src/shared/domain/models/Organization.js';
+import { Organization } from '../../../../../src/organizational-entities/domain/models/Organization.js';
 import { domainBuilder, expect } from '../../../../test-helper.js';
 
 describe('Unit | Serializer | organization-serializer', function () {
